Add files via upload
The files contain the EDI data which are used in the inversion of resistivity model, and the the resistivity model inverted by the inversion software (ModEM) for the manuscript entitled " Three-dimensional magnetotelluric imaging of the SE Gonghe Basin: implication for the orogenic uplift in the northeastern margin of the Tibetan plateau"

The EDI  package file (EDI.rar ) contain 117 sites of broadband MT data. The inverted resistivity model (Inverted Resistivity.rho) is of ModEM format.